mucho ej

+++ src/ej04.py
@@ -1,5 +1,5 @@
 
-cel = int(input("Dame los grados en Celsius: "))
+cel = float(input("Dame los grados en Celsius: "))
 
 temp = (cel * 9 / 5) + 32
